+++ packages/components/src/molecules/ValidationOverview/ValidationOverview.stories.args.tsx
@@ -1,6 +1,7 @@
 import {ValidationOverviewType} from './types';
 
 export const MainValidationOverviewArgs: ValidationOverviewType = {
+  status: 'loaded',
   height: 800,
   onProblemSelect: problem => {
     console.log('Problem:', problem);