+++ packages/validation/src/__tests__/MonokleValidator.test.ts
@@ -1,16 +1,12 @@
 import Ajv from "ajv";
 import { expect, it } from "vitest";
-import { ResourceParser } from "../common/resourceParser.js";
-import {
-  createDefaultMonokleValidator,
-  MonokleValidator,
-} from "../MonokleValidator.js";
-import { processRefs } from "../references/process.js";
+import { processRefs, ResourceParser,  createDefaultMonokleValidator, MonokleValidator, } from "../index.js";
+
 
 // Usage note: This library relies on fetch being on global scope!
 import "isomorphic-fetch";
 import { RESOURCES } from "./badResources.js";
-import { extractK8sResources, readDirectory } from "./testUtils";
+import { extractK8sResources, readDirectory } from "./testUtils.js";
 
 it("should be simple to configure", async () => {
   const parser = new ResourceParser();