+++ packages/validation/src/config/read.browser.ts
@@ -1,0 +1,7 @@
+import type { Config } from "./parse.js";
+
+export async function readConfig(
+  path: string = ''
+): Promise<Config | undefined> {
+  return undefined;
+}